# [Xamarin.Android.build.Tasks] AOT can now be used with Fast Deployment