--- a/.external
+++ b/.external
@@ -1,2 +1,2 @@
-xamarin/monodroid:master@1b907d680cc6561dcfaddc6f997d2f6ff5456644
+xamarin/monodroid:profiled-aot-debug-mode@8be5e209e2cfdc1bf689955e080476f16f3eddba
 mono/mono:2019-08@8946e49a974ea8b75fe5b8b7e93ffd4571521a85