Context: https://github.com/xamarin/monodroid/pull/1055
Fixes: https://github.com/xamarin/xamarin-android/issues/3828

Moved the `<Aot/>` MSBuild task call from `_BuildApkEmbed` to a new
`_AotAssemblies` target. This allows it to work for Fast Deployment
and the `_BuildApkFastDev` target *and* the `_BuildApkEmbed` target.

The other caveat is that we don't want to AOT the user's assemblies in
their project. We don't want AOT to run for small code changes, and so
we can AOT "framework assemblies" and any user assembly where
`%(ReferenceSourceTarget)` isn't a `<ProjectReference/>`:

    <_AssembliesToAot Condition=" '$(AndroidUseDebugRuntime)' != 'True' " Include="@(_ResolvedUserAssemblies);@(_ShrunkFrameworkAssemblies)" />
    <_AssembliesToAot Condition=" '$(AndroidUseDebugRuntime)' == 'True' " Include="@(_ShrunkFrameworkAssemblies)" />
    <_AssembliesToAot Condition=" '$(AndroidUseDebugRuntime)' == 'True' And '%(_ResolvedUserAssemblies.ReferenceSourceTarget)' != 'ProjectReference' " Include="@(_ResolvedUserAssemblies)" />

This way we aren't running the AOT compiler on any of the user's
assemblies. This would severely impact incremental build performance.

## Results ##

Results for `samples/HelloWorld`:

    Before:
    ActivityTaskManager: Displayed com.xamarin.android.helloworld/example.MainActivity: +2s449m
    After:
    ActivityTaskManager: Displayed com.xamarin.android.helloworld/example.MainActivity: +2s439ms

~10ms better? It really is a "Hello World".

Then I tried the Xamarin.Forms integration project:

    Before:
    ActivityTaskManager: Displayed Xamarin.Forms_Performance_Integration/xamarin.forms.performance.integration.MainActivity: +4s189ms
    After:
    ActivityTaskManager: Displayed Xamarin.Forms_Performance_Integration/xamarin.forms.performance.integration.MainActivity: +3s527ms

It seems about ~650ms better?

My comparison is:

    > git clean -dxf .\tests\
    > adb uninstall Xamarin.Forms_Performance_Integration
    > adb logcat -c
    > xabuild .\tests\Xamarin.Forms-Performance-Integration\Droid\Xamarin.Forms.Performance.Integration.Droid.csproj /r /t:Install,_Run /p:AndroidEnableProfiledAot=true

Versus:

    > xabuild .\tests\Xamarin.Forms-Performance-Integration\Droid\Xamarin.Forms.Performance.Integration.Droid.csproj /r /t:Install,_Run

It seems like AOT'd assemblies are loaded, too, I got logging such as:

    > adb logcat -d | grep 'Loaded AOT image'
    11-19 10:29:25.358 29065 29065 I monodroid: Loaded AOT image 'libaot-mscorlib.dll.so'
    11-19 10:29:25.368 29065 29065 I monodroid: Loaded AOT image 'libaot-FormsViewGroup.dll.so'
    11-19 10:29:25.373 29065 29065 I monodroid: Loaded AOT image 'libaot-Mono.Android.dll.so'
    11-19 10:29:25.379 29065 29065 I monodroid: Loaded AOT image 'libaot-System.Core.dll.so'
    11-19 10:29:25.381 29065 29065 I monodroid: Loaded AOT image 'libaot-System.dll.so'
    11-19 10:29:25.382 29065 29065 I monodroid: Loaded AOT image 'libaot-Mono.Security.dll.so'
    11-19 10:29:25.384 29065 29065 I monodroid: Loaded AOT image 'libaot-Java.Interop.dll.so'
    11-19 10:29:25.385 29065 29065 I monodroid: Loaded AOT image 'libaot-System.Net.Http.dll.so'
    11-19 10:29:25.387 29065 29065 I monodroid: Loaded AOT image 'libaot-System.Xml.dll.so'
    11-19 10:29:25.389 29065 29065 I monodroid: Loaded AOT image 'libaot-Newtonsoft.Json.dll.so'
    11-19 10:29:25.390 29065 29065 I monodroid: Loaded AOT image 'libaot-Plugin.Connectivity.Abstractions.dll.so'
    11-19 10:29:25.392 29065 29065 I monodroid: Loaded AOT image 'libaot-Plugin.Connectivity.dll.so'
    11-19 10:29:25.393 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Arch.Core.Common.dll.so'
    11-19 10:29:25.394 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Arch.Core.Runtime.dll.so'
    11-19 10:29:25.396 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Arch.Lifecycle.Common.dll.so'
    11-19 10:29:25.397 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Arch.Lifecycle.LiveData.Core.dll.so'
    11-19 10:29:25.398 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Arch.Lifecycle.LiveData.dll.so'
    11-19 10:29:25.399 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Arch.Lifecycle.Runtime.dll.so'
    11-19 10:29:25.401 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Arch.Lifecycle.ViewModel.dll.so'
    11-19 10:29:25.402 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.Animated.Vector.Drawable.dll.so'
    11-19 10:29:25.403 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.Annotations.dll.so'
    11-19 10:29:25.405 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.AsyncLayoutInflater.dll.so'
    11-19 10:29:25.406 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.Collections.dll.so'
    11-19 10:29:25.408 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.Compat.dll.so'
    11-19 10:29:25.409 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.VersionedParcelable.dll.so'
    11-19 10:29:25.410 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.CoordinaterLayout.dll.so'
    11-19 10:29:25.412 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.CustomView.dll.so'
    11-19 10:29:25.413 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.Core.UI.dll.so'
    11-19 10:29:25.414 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.ViewPager.dll.so'
    11-19 10:29:25.416 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.SwipeRefreshLayout.dll.so'
    11-19 10:29:25.417 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.Core.Utils.dll.so'
    11-19 10:29:25.419 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.CursorAdapter.dll.so'
    11-19 10:29:25.420 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.CustomTabs.dll.so'
    11-19 10:29:25.421 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.Design.dll.so'
    11-19 10:29:25.423 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.v7.CardView.dll.so'
    11-19 10:29:25.424 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.v7.AppCompat.dll.so'
    11-19 10:29:25.426 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.Fragment.dll.so'
    11-19 10:29:25.428 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.v7.RecyclerView.dll.so'
    11-19 10:29:25.429 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.Transition.dll.so'
    11-19 10:29:25.430 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.DocumentFile.dll.so'
    11-19 10:29:25.432 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.DrawerLayout.dll.so'
    11-19 10:29:25.433 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.Interpolator.dll.so'
    11-19 10:29:25.435 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.Loader.dll.so'
    11-19 10:29:25.436 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.LocalBroadcastManager.dll.so'
    11-19 10:29:25.437 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.Media.Compat.dll.so'
    11-19 10:29:25.439 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.Print.dll.so'
    11-19 10:29:25.440 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.SlidingPaneLayout.dll.so'
    11-19 10:29:25.442 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.v4.dll.so'
    11-19 10:29:25.443 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.v7.MediaRouter.dll.so'
    11-19 10:29:25.445 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.v7.Palette.dll.so'
    11-19 10:29:25.446 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Android.Support.Vector.Drawable.dll.so'
    11-19 10:29:25.448 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Forms.Core.dll.so'
    11-19 10:29:25.451 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Forms.Platform.Android.dll.so'
    11-19 10:29:25.452 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Forms.Platform.dll.so'
    11-19 10:29:25.453 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Forms.Xaml.dll.so'
    11-19 10:29:25.544 29065 29065 I monodroid: Loaded AOT image 'libaot-System.Runtime.Serialization.dll.so'
    11-19 10:29:25.545 29065 29065 I monodroid: Loaded AOT image 'libaot-System.ServiceModel.Internals.dll.so'
    11-19 10:29:25.622 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Forms.Performance.Integration.Droid.dll.so'
    11-19 10:29:25.642 29065 29065 I monodroid: Loaded AOT image 'libaot-netstandard.dll.so'
    11-19 10:29:25.794 29065 29065 I monodroid: Loaded AOT image 'libaot-System.Runtime.dll.so'
    11-19 10:29:26.142 29065 29065 I monodroid: Loaded AOT image 'libaot-System.Drawing.Common.dll.so'

None of the user's code is listed, only NuGet packages. :)