@jonathanpeppers 2019-11-19
> What is causing this? Are the nuget assemblies omitted or is the non-XF profile used?

The latest change I check `%(ReferenceSourceTarget)` as a way to include NuGet packages. But it looks like it is still picking up the head project:
```
11-19 10:29:25.622 29065 29065 I monodroid: Loaded AOT image 'libaot-Xamarin.Forms.Performance.Integration.Droid.dll.so'
```
Need to fix that.